Removed legacy backtesting, renamed files
