+++ u8timeseries/backtesting/__init__.py
@@ -3,5 +3,4 @@ Backtesting
 -----------
 """
 
-from .backtesting_udf import get_train_val_series, backtest_autoregressive_model
-from .forecasting_simulation import backtest_forecasting, backtest_regression
+from .backtesting import backtest_forecasting, backtest_regression
